requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
beniget==0.4.1
bleach==6.1.0
blinker==1.7.0
Brlapi==0.8.3
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
command-not-found==0.3
contourpy==1.2.0
cryptography==42.0.5
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
distro-info==1.1+ubuntu0.2
duplicity==2.2.3
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
extra-streamlit-components==0.1.71
Faker==24.11.0
fasteners==0.19
fastjsonschema==2.19.1
favicon==0.7.0
filelock==3.13.1
Flask==1.1.2
fonttools==4.50.0
fqdn==1.5.1
fs==2.4.16
future==1.0.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
gotrue==2.4.2
greenlet==3.0.3
gunicorn==19.9.0
h11==0.14.0
htbuilder==0.6.2
html5lib==1.1
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.22.2
ipython_genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==1.1.0
jaraco.classes==3.3.1
jaraco.context==4.3.0
jaraco.functools==4.0.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
keyring==25.0.0
kiwisolver==1.4.5
lab==8.1
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.1.0
lz4==4.3.3
macaroonbakery==1.3.4
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multipledispatch==1.0.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
netifaces==0.11.0
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
opencv-python==4.9.0.80
openpyxl==3.1.2
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
plotly==5.20.0
ply==3.11
postgrest==0.16.3
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==15.0.2
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
pymdown-extensions==10.7.1
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==3.1.2
pyRFC3339==1.1
pyrsistent==0.20.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.49
python-json-logger==2.0.7
pythran==0.15.0
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
realtime==1.0.4
referencing==0.34.0
reportlab==4.1.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.3
Send2Trash==1.8.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
speedtest-cli==2.1.3
SQLAlchemy==2.0.29
st-annotated-text==4.0.1
st-login-form==0.2.2
stack-data==0.6.3
storage3==0.7.4
streamlit==1.32.2
streamlit-authenticator==0.3.2
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-js-eval==0.1.7
streamlit-keyup==0.2.4
streamlit-option-menu==0.3.12
streamlit-shadcn-ui==0.1.18
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
StrEnum==0.4.15
supabase==2.4.3
supafunc==0.4.5
sympy==1.12
systemd-python==234
tenacity==8.2.3
terminado==0.18.1
testpath==0.6.0
threadpoolctl==3.4.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
traitlets==5.14.2
txt2tags==3.9
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==15.1.0
uri-template==1.3.0
urllib3==2.2.1
usb-creator==0.3.7
validators==0.24.0
virtualenv==20.25.1
wadllib==1.3.6
watchdog==4.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==1.0.1
widgetsnbextension==4.0.10
xdg==6.0.0
xkit==0.0.0
zipp==3.18.1
